+++ ec/__info__.py
@@ -1,0 +1,4 @@
+
+__author__ = 'Laufire Technologies'
+__email__ = 'laufiretechnologies@gmail.com'
+__version__ = '0.1.1'