+++ ec/__init__.py
@@ -1,6 +1,2 @@
-# -*- coding: utf-8 -*-
 
-__author__ = 'Laufire Technologies'
-__email__ = 'laufiretechnologies@gmail.com'
-__version__ = '0.1.0'
-
+from __info__ import __author__, __email__, __version__