+++ ec/types/basics.py
@@ -4,15 +4,21 @@ class yn(CustomType):
   """
   The classic y / n input that returns a truthy / falsy value.
   """
-  def __init__(self, desc=None):
-    CustomType.__init__(self, desc if desc is not None else 'y/n')
+  def __init__(self, desc=None, default=None):
+    label = desc if desc is not None else ('y/n%s' % ((' (%s)' % 'y' if default else 'n') if default is not None else ''))
+    
+    CustomType.__init__(self, label)
+    self.default = default
   
   def __call__(self, val):
     if val in 'yY':
-      return 1
+      return True
     
     elif val in 'nN':
-      return 0
+      return False
+      
+    elif self.default is not None:
+      return self.default
       
     else:
       raise ValueError('Invalid value.')