+++ scripts/examples/advanced/simple.py
@@ -1,7 +1,7 @@
 """Helper import for demonstrating advanced examples.
 """
 
-from ec.config import start, task, arg, group
+from ec.ec import start, task, arg, group
 
 @task
 @arg('arg1', type=int, desc= 'Value for arg1 (int)')